+++ godot_example/src/main.gd
@@ -58,10 +58,12 @@ func _on_get_authorization_status_button_pressed():
 func _on_get_authorization_status_success(result: RuStoreBillingUserAuthorizationStatus):
 	_loading.visible = false
 	OS.alert(str(result.authorized), "UserAuthorizationStatus")
+	result.free()
 
 func _on_get_authorization_status_failure(error: RuStoreError):
 	_loading.visible = false
 	_core_client.show_toast(error.description)
+	error.free()
 
 
 func _on_tab_container_tab_clicked(tab):
@@ -91,11 +93,13 @@ func _on_get_products_success(products: Array):
 func _on_get_products_failure(error: RuStoreError):
 	_loading.visible = false
 	_core_client.show_toast(error.description)
+	error.free()
 
 
 # Purchase product
 func _on_purchase_product_pressed(product: RuStoreProduct):
 	_billing_client.purchase_product(product.productId)
+	product.free()
 
 func _on_purchase_product_success(result: RuStorePaymentResult):
 	if result is RuStorePaymentResult.Success:
@@ -108,9 +112,11 @@ func _on_purchase_product_success(result: RuStorePaymentResult):
 		_core_client.show_toast("InvalidPaymentState")
 	else:
 		_core_client.show_toast("RuStorePaymentResult")
+	result.free()
 
 func _on_purchase_product_failure(error: RuStoreError):
 	_core_client.show_toast(error.description)
+	error.free()
 
 
 # Update purchases list
@@ -135,12 +141,14 @@ func _on_get_purchases_success(purchases: Array):
 func _on_get_purchases_failure(error: RuStoreError):
 	_loading.visible = false
 	_core_client.show_toast(error.description)
+	error.free()
 
 
 # Confirm purchase
 func _on_confirm_purchase_pressed(purchase: RuStorePurchase):
 	_loading.visible = true
 	_billing_client.confirm_purchase(purchase.purchaseId, purchase.developerPayload)
+	purchase.free()
 
 func _on_confirm_purchase_success(purchase_id: String):
 	_loading.visible = false
@@ -150,12 +158,14 @@ func _on_confirm_purchase_success(purchase_id: String):
 func _on_confirm_purchase_failure(purchase_id: String, error: RuStoreError):
 	_loading.visible = false
 	_core_client.show_toast(purchase_id + " " + error.description)
+	error.free()
 
 
 # Delete purchase
 func _on_delete_purchase_pressed(purchase: RuStorePurchase):
 	_loading.visible = true
 	_billing_client.delete_purchase(purchase.purchaseId)
+	purchase.free()
 
 func _on_delete_purchase_success(purchase_id: String):
 	_loading.visible = false
@@ -165,34 +175,43 @@ func _on_delete_purchase_success(purchase_id: String):
 func _on_delete_purchase_failure(purchase_id: String, error: RuStoreError):
 	_loading.visible = false
 	_core_client.show_toast(purchase_id + " " + error.description)
+	error.free()
 
 
 # Get purchase info
 func _on_get_purchase_info_pressed(purchase: RuStorePurchase):
 	_loading.visible = true
 	_billing_client.get_purchase_info(purchase.purchaseId)
+	purchase.free()
 
 func _on_get_purchase_info_success(purchase: RuStorePurchase):
 	_loading.visible = false
 	OS.alert(purchase.language + "\n" + purchase.amountLabel, purchase.productId)
+	purchase.free()
 
 func _on_get_purchase_info_failure(purchase_id: String, error: RuStoreError):
 	_loading.visible = false
 	_core_client.show_toast(purchase_id + " " + error.description)
+	error.free()
 
 
 # Debug logs
 func _on_payment_logger_debug(error: RuStoreError, message: String, tag: String):
 	_core_client.show_toast(tag + ": " + message)
+	error.free()
 
 func _on_payment_logger_error(error: RuStoreError, message: String, tag: String):
 	_core_client.show_toast(tag + ": " + message)
+	error.free()
 	
 func _on_payment_logger_info(error: RuStoreError, message: String, tag: String):
 	_core_client.show_toast(tag + ": " + message)
+	error.free()
 	
 func _on_payment_logger_verbose(error: RuStoreError, message: String, tag: String):
 	_core_client.show_toast(tag + ": " + message)
+	error.free()
 	
 func _on_payment_logger_warning(error: RuStoreError, message: String, tag: String):
 	_core_client.show_toast(tag + ": " + message)
+	error.free()